first pass converting manager code to use hashicat
All the tests but 2 are passing. Those 2, multipass and
remove_unused, tested functionality now in hashicat (and tested in
hashicat) and were hard to update, so I removed them for now.

Dedup has all been commented out/removed for now. Will add it back if
needed but I'm going to try to use this as an opportunity to remove it.

Lots of code is just commented out and needs to be reviewed to be sure
all the functionality is still there.